feat : replace duplicate things with mesh class
intialization and setup of the vao, vbo and ebo are already in the mesh class. But there is nothing for empty buffer and various buffer usage. I add more method about generating empty mesh and set buffer usage and then replace buffer handling code in the FluidBuffer class to the Mesh class. It makes FluidBuffer class code more simpler.